Add files via upload
This is a list of files related to web development training in front-end languages such as HTML, CSS, and JavaScript. In this list, you will be able to see what I am doing so that you or I can create a beautiful, attractive website with a good appearance. However, this project mainly concerns an introduction for beginners in front-end development, but it is also useful for experts.